+++ components/SortMenu.js
@@ -1,53 +1,33 @@
 import React from "react"
 import Link from "next/link"
 
-import database from "../database.json"
 import FaArrowUp from "react-icons/lib/fa/arrow-up"
 import FaArrowDown from "react-icons/lib/fa/arrow-down"
 import FaStar from "react-icons/lib/fa/star"
 import FaDownload from "react-icons/lib/fa/download"
-import Dropdown from "./Dropdown/Dropdown"
 
 const SortMenu = () => (
-  <div className="sortNav">
-    <Dropdown categories={database.categories.map(x => x)} />
-    <ul>
-      <div className="header">Sort By</div>
-      <li>
-        <label>
-          <input checked type="checkbox" />
-          NPM
-        </label>
-        <div className="icon">
-          <FaDownload />
-        </div>
-      </li>
-      <li>
-        <label>
-          <input checked type="checkbox" />
-          GitHub
-        </label>
-        <div className="icon">
-          <FaStar />
-        </div>
-      </li>
-    </ul>
+  <ul>
+    <div className="header">Sort By</div>
+    <li>
+      <label>
+        <input checked type="checkbox" />
+        NPM
+      </label>
+      <div className="icon">
+        <FaDownload />
+      </div>
+    </li>
+    <li>
+      <label>
+        <input checked type="checkbox" />
+        GitHub
+      </label>
+      <div className="icon">
+        <FaStar />
+      </div>
+    </li>
     <style jsx>{`
-      .sortNav {
-        grid-row: 2;
-        grid-column: col-start / span 2;
-        font-size: 1rem;
-        font-weight: 300;
-        top: 25px;
-        position: -webkit-sticky;
-        position: sticky;
-        box-sizing: border-box;
-        height: 80vh;
-
-        display: flex;
-        flex-direction: column;
-      }
-
       input[type="checkbox"] {
         border: solid;
         border-radius: 4px;
@@ -75,7 +55,7 @@ const SortMenu = () => (
         width: 80%;
       }
     `}</style>
-  </div>
+  </ul>
 )
 
 export default SortMenu
